+++ packages/tokit/src/utils/modify-package-json.ts
@@ -1,13 +1,7 @@
 import { writeFileSync } from 'fs'
 
 import { json, pathOf, prettierString } from '@toktokhan-dev/node'
-import {
-  DataOrFn,
-  Obj,
-  awaitted,
-  pass,
-  runIfFn,
-} from '@toktokhan-dev/universal'
+import { DataOrFn, Obj, awaited, pass, runIfFn } from '@toktokhan-dev/universal'
 
 import { flow, merge } from 'lodash'
 
@@ -24,5 +18,5 @@ export const modifyPackageJson = (data: DataOrFn<Obj>, path: string) =>
     updator(data),
     JSON.stringify,
     (src) => prettierString(src, { parser: 'json' }),
-    awaitted((src) => writeFileSync(pathOf('package.json', path), src)),
+    awaited((src) => writeFileSync(pathOf('package.json', path), src)),
   )()