+++ packages/tokit/src/utils/proceed-git.ts
@@ -1,5 +1,5 @@
 import { $ } from '@toktokhan-dev/node'
-import { awaitted } from '@toktokhan-dev/universal'
+import { awaited } from '@toktokhan-dev/universal'
 
 import { flow } from 'lodash'
 import simpleGit from 'simple-git'
@@ -15,19 +15,19 @@ export const proceedGit = async ({ cloneUrl, baseDir }: RemoteGitParams) => {
   try {
     await flow(
       async () => $('rm', ['-rf', '.git']),
-      // awaitted(infoLog('Remove .git')),
-      awaitted(() => git.init()),
-      // awaitted(infoLog('Initialized git repository')),
-      awaitted(() => git.add('.')),
-      // awaitted(infoLog('Added all files')),
-      awaitted(() => git.commit('Upload TOKIT`s template')),
-      // awaitted(infoLog('Committed changes')),
-      awaitted(() => git.branch(['-M', 'main'])),
-      // awaitted(infoLog('Renamed branch to main')),
-      awaitted(() => git.addRemote('origin', cloneUrl)),
-      // awaitted(infoLog('Added remote origin')),
-      awaitted(() => git.push(['-u', 'origin', 'main'])),
-      // awaitted(infoLog('Pushed to origin/main')),
+      // awaited(infoLog('Remove .git')),
+      awaited(() => git.init()),
+      // awaited(infoLog('Initialized git repository')),
+      awaited(() => git.add('.')),
+      // awaited(infoLog('Added all files')),
+      awaited(() => git.commit('Upload TOKIT`s template')),
+      // awaited(infoLog('Committed changes')),
+      awaited(() => git.branch(['-M', 'main'])),
+      // awaited(infoLog('Renamed branch to main')),
+      awaited(() => git.addRemote('origin', cloneUrl)),
+      // awaited(infoLog('Added remote origin')),
+      awaited(() => git.push(['-u', 'origin', 'main'])),
+      // awaited(infoLog('Pushed to origin/main')),
     )()
   } catch (error) {
     console.error('Error git process:', error)